+++ chat/ConversationView.vue
@@ -135,9 +135,12 @@
         <bbcode-editor v-model="conversation.enteredText" @keydown="onKeyDown" :extras="extraButtons" @input="keepScroll"
             :classes="'form-control chat-text-box' + (isChannel(conversation) && conversation.isSendingAds ? ' ads-text-box' : '')"
             :hasToolbar="settings.bbCodeBar" ref="textBox" style="position:relative;margin-top:5px"
-            :maxlength="isChannel(conversation) || isPrivate(conversation) ? conversation.maxMessageLength : undefined">
+            :maxlength="isChannel(conversation) || isPrivate(conversation) ? conversation.maxMessageLength : undefined"
+            :characterName="ownName"
+            >
+
             <span v-if="isPrivate(conversation) && conversation.typingStatus !== 'clear'" class="chat-info-text">
-                {{l('chat.typing.' + conversation.typingStatus, conversation.name)}}
+              <user :character="conversation.character" :match="false" :bookmark="false"></user>&nbsp;{{l('chat.typing.' + conversation.typingStatus, '').trim()}}
             </span>
             <div v-show="conversation.infoText" class="chat-info-text">
                 <span class="fa fa-times" style="cursor:pointer" @click.stop="conversation.infoText = ''"></span>
@@ -203,13 +206,22 @@
     import CharacterChannelList from './character/CharacterChannelList.vue';
     import * as _ from 'lodash';
     import Dropdown from '../components/Dropdown.vue';
-
+    import { EventBus } from './preview/event-bus';
 
     @Component({
         components: {
-            user: UserView, 'bbcode-editor': Editor, 'manage-channel': ManageChannel, settings: ConversationSettings,
-            logs: Logs, 'message-view': MessageView, bbcode: BBCodeView(core.bbCodeParser), 'command-help': CommandHelp,
-            'ad-view': CharacterAdView, 'channel-list': CharacterChannelList, dropdown: Dropdown, adSettings: ConversationAdSettings
+            user: UserView,
+            'bbcode-editor': Editor,
+            'manage-channel': ManageChannel,
+            settings: ConversationSettings,
+            logs: Logs,
+            'message-view': MessageView,
+            bbcode: BBCodeView(core.bbCodeParser),
+            'command-help': CommandHelp,
+            'ad-view': CharacterAdView,
+            'channel-list': CharacterChannelList,
+            dropdown: Dropdown,
+            adSettings: ConversationAdSettings
         }
     })
     export default class ConversationView extends Vue {
@@ -246,15 +258,20 @@
         isPrivate = Conversation.isPrivate;
         showNonMatchingAds = true;
 
+        ownName?: string;
 
         @Hook('beforeMount')
         async onBeforeMount(): Promise<void> {
+          this.updateOwnName();
+
           this.showNonMatchingAds = !await core.settingsStore.get('hideNonMatchingAds');
         }
 
 
         @Hook('mounted')
         mounted(): void {
+            this.updateOwnName();
+
             this.extraButtons = [{
                 title: 'Help\n\nClick this button for a quick overview of slash commands.',
                 tag: '?',
@@ -295,7 +312,12 @@
 
             this.$watch(() => this.conversation.adManager.isActive(), () => (this.refreshAutoPostingTimer()));
             this.refreshAutoPostingTimer();
+
+            this.configUpdateHook = () => this.updateOwnName();
+            EventBus.$on('configuration-update', this.configUpdateHook);
         }
+
+        protected configUpdateHook: any;
 
         @Hook('destroyed')
         destroyed(): void {
@@ -305,11 +327,17 @@
             clearInterval(this.searchTimer);
             clearInterval(this.autoPostingUpdater);
             clearInterval(this.adCountdown);
+
+            EventBus.$off('configuration-update', this.configUpdateHook);
         }
 
         hideSearch(): void {
             this.showSearch = false;
             this.searchInput = '';
+        }
+
+        updateOwnName(): void {
+            this.ownName = core.state.settings.risingShowPortraitNearInput ? core.characters.ownCharacter?.name : undefined;
         }
 
         get conversation(): Conversation {
@@ -328,6 +356,8 @@
 
         @Watch('conversation')
         conversationChanged(): void {
+            this.updateOwnName();
+
             if(!anyDialogsShown) (<Editor>this.$refs['textBox']).focus();
             this.$nextTick(() => setTimeout(() => this.messageView.scrollTop = this.messageView.scrollHeight));
             this.scrolledDown = true;